feat: add new Teams admin panel section and remove old members page
