+++ components/admin-panel/AdminPanelSection.js
@@ -19,6 +19,7 @@ import FinancialContributions from './sections/FinancialContributions';
 import HostVirtualCards from './sections/HostVirtualCards';
 import InvoicesReceipts from './sections/InvoicesReceipts';
 import NotificationsSettings from './sections/NotificationsSettings';
+import TeamSettings from './sections/Team';
 import {
   COLLECTIVE_SECTIONS,
   FISCAL_HOST_SECTIONS,
@@ -35,6 +36,7 @@ const ADMIN_PANEL_SECTIONS = {
   [HOST_DASHBOARD_SECTIONS.REPORTS]: HostDashboardReports,
   [HOST_DASHBOARD_SECTIONS.HOST_VIRTUAL_CARDS]: HostVirtualCards,
   [COLLECTIVE_SECTIONS.NOTIFICATIONS]: NotificationsSettings,
+  [COLLECTIVE_SECTIONS.TEAM]: TeamSettings,
 };
 
 const FISCAL_HOST_SETTINGS_SECTIONS = {
@@ -59,7 +61,6 @@ const AdminPanelSection = ({ collective, isLoading, section, subpath }) => {
     );
   }
 
-  // Host Dashboard
   const AdminSectionComponent = ADMIN_PANEL_SECTIONS[section];
   if (AdminSectionComponent) {
     return (
