+++ components/admin-panel/types.ts
@@ -1,0 +1,7 @@
+import type { Account } from '../../lib/graphql/types/v2/graphql';
+
+export type AdminSectionProps = {
+  account: Partial<Account>;
+  subpath: string;
+  hostSlug: string;
+};
