+++ components/collective-page/CollectiveNotificationBar.js
@@ -182,7 +182,7 @@ const getNotification = (intl, status, collective, host, LoggedInUser, refetch) 
       }),
       type: 'warning',
       actions: (
-        <NotificationBarLink href={`/${collective.slug}/admin/members`}>
+        <NotificationBarLink href={`/${collective.slug}/admin/team`}>
           <FormattedMessage defaultMessage="Manage members" />
         </NotificationBarLink>
       ),
